fix(dtype): eliminate silent float64 upcasts in scale-and-rotate paths (#72)
* fix(dtype): eliminate silent float64 upcasts in scale-and-rotate paths

Several hot-path expressions multiplied a float32 array by np.sqrt(int)
or divided it by a Python literal, both of which produce a numpy
float64 scalar and promote the result to float64.  In most call sites
an explicit .astype(np.float32) tapped the value back to float32, but
one site (_index.py:285) had no astype at all -- the variable was
annotated NDArray[np.float32] while its runtime dtype was float64,
which then propagated through np.searchsorted on the boundary table.

Fix the root cause at each site by wrapping the scalar in
np.float32(...) so the multiplication stays in float32 throughout,
then drop the redundant astype copies that were compensating for the
now-absent upcast.

Sites touched (runtime dtype before vs after):

  _rotation.py:95-103   y / np.sqrt(d)          float64 -> float32
  _index.py:285         rotated * np.sqrt(pdim) float64 -> float32  (latent bug)
  _index.py:427         q_rot   * np.sqrt(pdim) float64 (tapped by astype) -> float32
  _residual.py:163,233  rotated * np.sqrt(pdim) float64 (tapped) -> float32

Additionally remove .astype(np.float32) copies that were pure
identity on float32 (matmul of two float32 operands, slice of a
float32 array, broadcast division by float32):

  _pq.py:194,195,215,217
  _ivfpq.py:258,261,262,279,281,967,969
  _index.py:462,463,488,489,508

Preserve the per-subspace contiguity copy in IVFPQ fit() -- the old
.astype() on residuals[:, slice] was doing double duty (dtype guard
AND making a non-contiguous view contiguous for BLAS).  Replace with
an explicit np.ascontiguousarray() so the intent is legible.

Scope disclosure: end-to-end fit() and search_batch() latency on an
N=50k M=96 synthetic FIQA-shaped corpus is unchanged within noise
(<= 2%).  The value here is correctness (matching annotations to
runtime dtype, removing latent-float64 paths) and code hygiene
(fewer 'is this astype a dtype guard or a memory copy?' questions
for a reader), not measurable speedup.  Shipped anyway because the
latent upcast in _index.py:285 was a real type/annotation
divergence that would have bitten us eventually.

All 199 tests pass, mypy --strict clean, ruff clean.

* fix(dtype): address PR #72 bot review -- 4 missed upcasts + regression tests

Gemini and Copilot both caught real regressions I introduced in the
earlier commit.  Ironic given the stated goal was to eliminate silent
upcasts, exactly validates the reason bots review PRs.

Gemini found four sites where a Python-literal or Python-float
divisor was upcasting a float32 array to float64 under pre-NEP-50
numpy (>= 1.24 in pyproject), now not tapped back by any astype:

- _ivfpq.py:257  rot /= einsum + 1e-12
- _ivfpq.py:271  q / q_norm  (q_norm = float(np.linalg.norm(q)))
- _ivfpq.py:983  q_pre_all /= einsum + 1e-12  (search_batch path)
- _pq.py:190     rot /= einsum + 1e-12
- _pq.py:207     q / q_norm
- _residual.py:229  q / q_norm
- _index.py:427  q / q_norm

Fix the root cause by wrapping each scalar in np.float32():
  np.float32(1e-12), np.float32(q_norm), np.float32(np.linalg.norm(rot))

Copilot suggested regression tests for these dtype invariants.
Added tests/test_dtype_invariants.py with 5 checks:

- rht() returns float32 for float32 input (1-D and 2-D, across all
  supported pdim values)
- SnapIndex.add_batch stores float32 _norms
- PQSnapIndex._preprocess + _preprocess_single return float32 under
  all three branches (plain, use_rht, use_opq)
- IVFPQSnapIndex._preprocess + _preprocess_single + the separate
  search_batch normalisation path all stay float32 under the same
  three branches
- ResidualSnapIndex search path keeps stored _norms float32

All 204 tests pass (199 existing + 5 new), mypy --strict clean, ruff
clean.  Two cast() additions needed on returns of 'q / np.float32(q_norm)'
because numpy stubs do not infer the scalar wrap correctly.

* fix(dtype): address self-review on PR #72 -- 2 blockers + test discriminators

Ran an independent code-reviewer subagent pass after the bot reviews
(Gemini + Copilot) in case something slipped through.  It found two
real blockers plus methodology issues the bots did not catch.

BLOCKER 1 -- _apply_qjl_arrays silently upcasts to float64.
_index.py:540, 544:

  q_unit_rot: NDArray[np.float32] = q_scaled / np.sqrt(self._pdim)
  correction: NDArray[np.float32] = (
      np.sqrt(np.pi / 2.0) / self._pdim * rnorms * qjl_dots
  )

Both annotations were lying: np.sqrt(int) and np.sqrt(float) return
numpy float64 scalars that upcast their float32 operands under every
supported numpy version.  Result: the whole use_prod=True correction
path silently ran in float64, masked only by float(scores[i]) at the
return boundary of search().  Fix: wrap each scalar in np.float32().

BLOCKER 2 -- the SnapIndex dtype test did not discriminate.
tests/test_dtype_invariants.py.  The original test only asserted
idx._norms.dtype == np.float32, but _norms is the input-norm capture
(always float32 via the cast at _index.py:276) regardless of the
surrounding bug.  The test would have passed on the pre-PR buggy
code.  Rewrote to assert:

- idx._rnorms.dtype == np.float32 (the output of the fixed
  residual_norms expression; genuinely float64 under the buggy code)
- _apply_qjl_arrays(...) output dtype (direct call to the path fixed
  in this commit; verified to FAIL with 'float64' on the stashed
  pre-fix _index.py)

Also strengthened the IVFPQ batch test (S1): monkey-patches
snapvec._ivfpq.rht to capture the input dtype under
search_batch(..., use_rht=True), asserting it is float32.  The
rht() spy fires on every query and catches a regression at
_ivfpq.py:983 directly instead of via a downstream comparison.

Bench fix (S2) -- the perf bench now covers the RHT path it was
supposed to measure.  Added:

- bench_fit(use_rht=True) variant so the rotation-normalisation
  expression runs at all
- bench_search_batch(use_rht=True) variant for the batch path
- bench_snap_add_batch with use_prod=True to cover the residual
  PQ scale/rotate path in SnapIndex that the IVFPQ benches miss
- Warmup now uses IVFPQSnapIndex (same code path) instead of
  PQSnapIndex (different k-means pattern, no IVF coarse stage)

All 204 tests pass, mypy --strict clean, ruff clean.

---------

+++ snapvec/_residual.py
@@ -154,13 +154,13 @@ class ResidualSnapIndex(FreezableIndex):
                 np.where(raw_norms > 1e-10, raw_norms, np.float32(1.0)),
             )
             units = cast("NDArray[np.float32]", arr / safe[:, None])
-            batch_norms = np.where(raw_norms > 1e-10, raw_norms, 0.0).astype(np.float32)
+            batch_norms = np.where(raw_norms > 1e-10, raw_norms, np.float32(0.0))
 
         pdim = self._pdim
         padded = np.zeros((n, pdim), dtype=np.float32)
         padded[:, : self.dim] = units
         rotated = rht(padded, self.seed)
-        scaled = (rotated * np.sqrt(pdim)).astype(np.float32)
+        scaled = rotated * np.float32(np.sqrt(pdim))
 
         c1 = np.clip(
             np.searchsorted(self._thr1, scaled),
@@ -226,11 +226,13 @@ class ResidualSnapIndex(FreezableIndex):
         if q_norm < 1e-10:
             return []
         pdim = self._pdim
-        q_unit = q / q_norm
+        # Python float ``q_norm`` is float64 under pre-NEP-50 numpy; wrap
+        # so ``q_unit`` stays in float32.
+        q_unit = q / np.float32(q_norm)
         q_padded = np.zeros(pdim, dtype=np.float32)
         q_padded[: self.dim] = q_unit
         q_rot = rht(q_padded[None, :], self.seed)[0]
-        q_scaled = (q_rot * np.sqrt(pdim)).astype(np.float32)
+        q_scaled = q_rot * np.float32(np.sqrt(pdim))
 
         if rerank_M is None:
             decoded = self._cent1[self._codes1] + self._sigma_r * self._cent2[self._codes2]
